Satisfy clang even more

+++ joint_trajectory_controller/include/joint_trajectory_controller/tolerances.hpp
@@ -199,7 +199,7 @@ SegmentTolerances get_segment_tolerances(
           .c_str());
       return default_tolerances;
     }
-    auto i = std::distance(joints.cbegin(), it);
+    auto i = static_cast<size_t>(std::distance(joints.cbegin(), it));
     std::string interface = "";
     try
     {
